code/requirements.txt
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.1
comm==0.2.2
contourpy
coverage==7.8.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
geographiclib==2.0
geopy==2.4.1
haversine==2.9.0
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
jedi==0.19.2
Jinja2==3.1.6
matplotlib
mpmath==1.3.0
nest-asyncio==1.6.0
networkx
numpy
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
pprintpp==0.4.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycountry==24.6.1
pycountry-convert==0.7.2
Pygments==2.19.1
pyparsing==3.2.3
pypng==0.20220715.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
repoze.lru==0.7
requests==2.32.3
reverse_geocoder==1.5.1
safetensors==0.5.3
scipy
seaborn==0.13.2
setuptools==80.1.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tinytag==2.1.1
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
wheel==0.45.1
